1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
-* array_iterator - executes a function given as a parameter on each element of an array.
+* array_iterator - executes a function given as a parameter.
 * @array: array of elements
 * @size: the size of the array
 * @action: a pointer to the function you need to use
@@ -10,7 +10,7 @@
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
 int x;
-long unsigned int i;
+unsigned int i;
 if (!array || !action)
 return;
